+++ concordium-consensus/src/Concordium/Scheduler/Environment.hs
@@ -154,10 +154,6 @@ class (Monad m, StaticInformation m, CanRecordFootprint (Footprint (ATIStorage m
 
   -- |Convert the given energy amount into an amount of GTU. The exchange
   -- rate can vary depending on the current state of the blockchain.
-  -- TODO: In this setup the exchange rate is determined by the blockchain, and
-  -- the user (aka sender of the transaction) cannot choose to pay more to have
-  -- their transaction prioritised. If the user can choose to do so then this
-  -- function needs to be replaced.
   energyToGtu :: Energy -> m Amount
 
   -- *Operations related to bakers.