+++ concordium-node/src/concordium_dns/dns.rs
@@ -6,6 +6,8 @@ const DNS_ANCHOR_2: &str = ". IN DNSKEY 256 3 8 AwEAAYvxrQOOujKdZz+37P+oL4l7e35/
 const DNS_ANCHOR_3: &str = ". IN DNSKEY 257 3 8 AwEAAagAIKlVZrpC6Ia7gEzahOR+9W29euxhJhVVLOyQbSEW0O8gcCjFFVQUTf6v58fLjwBd0YI0EzrAcQqBGCzh/RStIoO8g0NfnfL2MTJRkxoXbfDaUeVPQuYEhg37NZWAJQ9VnMVDxP/VHL496M/QZxkjf5/Efucp2gaDX6RS6CXpoY68LsvPVjR0ZSwzz1apAzvN9dlzEheX7ICJBBtuA6G3LQpzW5hOA2hzCTMjJPJ8LbqF6dsV6DoBQzgul0sGIcGOYl7OyQdXfZ57relSQageu+ipAdTTJ25AsRTAoub8ONGcLmqrAmRLKBP1dfwhYB4N7knNnulqQxA+Uk1ihz0=";
 
 #[derive(Copy, Clone, Debug)]
+// We disable this lint for this specific type since A, AAAA, and TXT are standard acronyms.
+#[allow(clippy::upper_case_acronyms)]
 enum LookupType {
     A    = 1,
     AAAA = 28,