+++ concordium-node/src/consensus_ffi/helpers.rs
@@ -101,8 +101,8 @@ impl TryFrom<String> for HashBytes {
     fn try_from(value: String) -> Result<Self, Self::Error> { Self::from_str(value.as_str()) }
 }
 
-impl Into<String> for HashBytes {
-    fn into(self) -> String { self.to_string() }
+impl From<HashBytes> for String {
+    fn from(x: HashBytes) -> Self { x.to_string() }
 }
 
 // a short, 8-character beginning of the SHA
@@ -209,61 +209,55 @@ impl ConsensusFfiResponse {
     pub fn is_successful(self) -> bool {
         use ConsensusFfiResponse::*;
 
-        match self {
-            Success | PendingBlock | PendingFinalization | Asynchronous => true,
-            _ => false,
-        }
+        matches!(self, Success | PendingBlock | PendingFinalization | Asynchronous)
     }
 
     pub fn is_pending(self) -> bool {
         use ConsensusFfiResponse::*;
 
-        match self {
-            PendingBlock | PendingFinalization => true,
-            _ => false,
-        }
+        matches!(self, PendingBlock | PendingFinalization)
     }
 
     pub fn is_acceptable(self) -> bool {
         use ConsensusFfiResponse::*;
 
-        match self {
+        !matches!(
+            self,
             BakerNotFound
-            | DeserializationError
-            | InvalidResult
-            | Unverifiable
-            | BlockTooEarly
-            | ExpiryTooLate
-            | VerificationFailed
-            | NonexistingSenderAccount
-            | DuplicateNonce
-            | NonceTooLarge
-            | TooLowEnergy
-            | ConsensusShutDown => false,
-            _ => true,
-        }
+                | DeserializationError
+                | InvalidResult
+                | Unverifiable
+                | BlockTooEarly
+                | ExpiryTooLate
+                | VerificationFailed
+                | NonexistingSenderAccount
+                | DuplicateNonce
+                | NonceTooLarge
+                | TooLowEnergy
+                | ConsensusShutDown
+        )
     }
 
     pub fn is_rebroadcastable(self) -> bool {
         use ConsensusFfiResponse::*;
 
-        match self {
+        !matches!(
+            self,
             DeserializationError
-            | InvalidResult
-            | Unverifiable
-            | DuplicateEntry
-            | Stale
-            | IncorrectFinalizationSession
-            | BlockTooEarly
-            | ExpiryTooLate
-            | VerificationFailed
-            | NonexistingSenderAccount
-            | DuplicateNonce
-            | NonceTooLarge
-            | TooLowEnergy
-            | ConsensusShutDown => false,
-            _ => true,
-        }
+                | InvalidResult
+                | Unverifiable
+                | DuplicateEntry
+                | Stale
+                | IncorrectFinalizationSession
+                | BlockTooEarly
+                | ExpiryTooLate
+                | VerificationFailed
+                | NonexistingSenderAccount
+                | DuplicateNonce
+                | NonceTooLarge
+                | TooLowEnergy
+                | ConsensusShutDown
+        )
     }
 }